+++ app/models/micropost.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# .nodoc
+class Micropost < ApplicationRecord
+  belongs_to :user
+
+  validates :content, length: { maximum: 140 }, presence: true
+end