+++ app/models/user.rb
@@ -1,2 +1,9 @@
+# frozen_string_literal: true
+
+# .nodoc
 class User < ApplicationRecord
+  has_many :microposts
+
+  validates :email, presence: true
+  validates :name, presence: true
 end